Restore the image's previous experiment data when analyzing it
Restore the data/scripts generated by previous experiments for the target
image/re-hosting system. Pandawan copies out the output of the analysis
for each system and image to a dedicated directory every time we
analyze the image with a specific re-hosting system.

That happens, since all the supported re-hosting systems (e.g., FirmAE,
Firmadyne, FirmSolo and Pandawan) output their experiment results to
the same directories (e.g., scratch). Then, when the experiments are
complete the results are copied out to the dedicated directories for
each system.

So to be able to re-run (partially) the experiments with a re-hosting
system, restore the previous experiment results to the common (initial)
directories.